Restructure and reorganization of major scripts and code to reduced model
